Download updates (#546)
* Add rename download attempt from preview

* clean up renew cache some to help with readability

* Download on WiFi regardless of metered status  (From Tachiyomi)

* no reason to show md2 beta message for new debug installs

* update clean up downloads to be a little faster and hit some edge cases not thought of before

* fix bug where manga name does not match the file name and was deleting manga downloads of manga with special characters

+++ app/src/main/java/eu/kanade/tachiyomi/data/download/DownloadManager.kt
@@ -11,6 +11,7 @@ import eu.kanade.tachiyomi.source.Source
 import eu.kanade.tachiyomi.source.SourceManager
 import eu.kanade.tachiyomi.source.model.Page
 import rx.Observable
+import timber.log.Timber
 import uy.kohesive.injekt.injectLazy
 
 /**
@@ -253,25 +254,52 @@ class DownloadManager(val context: Context) {
     }
 
     /**
+     * return the list of all manga folders
+     */
+    fun getMangaFolders(source: Source): List<UniFile> {
+        return provider.findSourceDir(source)?.listFiles()?.toList() ?: emptyList()
+    }
+
+    /**
      * Deletes the directories of chapters that were read or have no match
      *
      * @param chapters the list of chapters to delete.
      * @param manga the manga of the chapters.
      * @param source the source of the chapters.
      */
-    fun cleanupChapters(allChapters: List<Chapter>, manga: Manga, source: Source): Int {
+    fun cleanupChapters(allChapters: List<Chapter>, manga: Manga, source: Source, removeRead: Boolean, removeNonFavorite: Boolean): Int {
         var cleaned = 0
+
+        if (removeNonFavorite && !manga.favorite) {
+            val mangaFolder = provider.getMangaDir(manga, source)
+            cleaned += 1 + (mangaFolder.listFiles()?.size ?: 0)
+            mangaFolder.delete()
+            cache.removeManga(manga)
+            return cleaned
+        }
+
         val filesWithNoChapter = provider.findUnmatchedChapterDirs(allChapters, manga, source)
         cleaned += filesWithNoChapter.size
         cache.removeFolders(filesWithNoChapter.mapNotNull { it.name }, manga)
         filesWithNoChapter.forEach { it.delete() }
-        val readChapters = allChapters.filter { it.read }
-        val readChapterDirs = provider.findChapterDirs(readChapters, manga, source)
-        readChapterDirs.forEach { it.delete() }
-        cleaned += readChapterDirs.size
-        cache.removeChapters(readChapters, manga)
+
+        if (removeRead) {
+            val readChapters = allChapters.filter { it.read }
+            val readChapterDirs = provider.findChapterDirs(readChapters, manga, source)
+            readChapterDirs.forEach { it.delete() }
+            cleaned += readChapterDirs.size
+            cache.removeChapters(readChapters, manga)
+        }
+
         if (cache.getDownloadCount(manga) == 0) {
-            provider.findChapterDirs(allChapters, manga, source).firstOrNull()?.parentFile?.delete() // Delete manga directory if empty
+            val mangaFolder = provider.getMangaDir(manga, source)
+            val size = mangaFolder.listFiles()?.size ?: 0
+            if (size == 0) {
+                mangaFolder.delete()
+                cache.removeManga(manga)
+            } else {
+                Timber.e("Cache and download folder doesn't match for %s", manga.title)
+            }
         }
         return cleaned
     }
@@ -307,6 +335,27 @@ class DownloadManager(val context: Context) {
         for ((manga, chapters) in pendingChapters) {
             val source = sourceManager.get(manga.source) ?: continue
             deleteChapters(chapters, manga, source)
+        }
+    }
+
+    /**
+     * Renames an already downloaded chapter
+     *
+     * @param manga the manga of the chapter.
+     * @param oldChapter the existing chapter with the old name.
+     * @param newChapter the target chapter with the new name.
+     */
+    fun renameChapter(source: Source, manga: Manga, oldChapter: Chapter, newChapter: Chapter) {
+        val oldName = provider.getChapterDirName(oldChapter)
+        val newName = provider.getChapterDirName(newChapter)
+        val mangaDir = provider.getMangaDir(manga, source)
+
+        val oldFolder = mangaDir.findFile(oldName)
+        if (oldFolder?.renameTo(newName) == true) {
+            cache.removeChapters(listOf(oldChapter), manga)
+            cache.addChapter(newName, mangaDir, manga)
+        } else {
+            Timber.e("Could not rename downloaded chapter: %s.", oldName)
         }
     }
